refactor: вынес утилиты getBaseUrl и getComponentSchemaUrl в module-utils
