+++ packages/core/components/microfrontend/import-remote/import-remote.ts
@@ -1,17 +1,15 @@
-import { encodePackageName } from '@colibrijs/module-utils';
+import { encodePackageName, getBaseUrl } from '@colibrijs/module-utils';
 import type { IComponent } from '@colibrijs/types';
 import { importRemote as importRemoteCsr } from '@module-federation/utilities';
 
-import { getBaseUrl } from './get-base-url';
 import { importRemoteSsr } from './import-remote-ssr';
 
 export function importRemote<T>(component: IComponent): Promise<T> {
-  const fullUrl = getBaseUrl(component);
   const ssr = typeof window === 'undefined';
 
   if (ssr) {
     return importRemoteSsr({
-      url: fullUrl,
+      url: getBaseUrl(component),
       scope: encodePackageName(component.libraryName),
       module: './component/',
       remoteEntryFileName: 'remote.server.js',
@@ -19,7 +17,7 @@ export function importRemote<T>(component: IComponent): Promise<T> {
   }
 
   return importRemoteCsr<T>({
-    url: fullUrl,
+    url: getBaseUrl(component),
     scope: encodePackageName(component.libraryName),
     module: './component/',
     remoteEntryFileName: 'remote.client.js',
