+++ packages/core/components/microfrontend/import-remote/index.ts
@@ -1,2 +1,1 @@
-export { getBaseUrl } from './get-base-url';
 export { importRemote } from './import-remote';
