+++ packages/module-utils/index.ts
@@ -1,1 +1,3 @@
 export { encodePackageName } from './encode-package-name';
+export { getBaseUrl } from './get-base-url';
+export { getComponentSchemaUrl } from './get-component-schema-url';
